requirements.txt
pandasai
pandas
langchain-groq
python-dotenv
streamlit
